pyproject.toml
[project]
name = "wingspan-agent"
version = "0.1.0"
description = "AI Agent using Google's Agent Development Kit (ADK)"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=0.2.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
]
